+++ ui/src/features/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './use-modal'
 export * from './use-oauth'
+export * from './use-outside-click'
