+++ ui/src/features/hooks/use-outside-click.ts
@@ -1,0 +1,16 @@
+import { RefObject, useEffect } from 'react'
+
+export function useOutsideClick(ref: RefObject<HTMLDivElement>, callback: () => void) {
+	useEffect(() => {
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		function handleClickOutside(event: any) {
+			if (ref.current && !ref.current.contains(event.target)) {
+				callback()
+			}
+		}
+		document.addEventListener('mousedown', handleClickOutside)
+		return () => {
+			document.removeEventListener('mousedown', handleClickOutside)
+		}
+	}, [callback, ref])
+}
